requirements.txt
# Core Dependencies
numpy==1.24.1
pandas==1.5.3
PyYAML==6.0.1
joblib==1.2.0
six==1.17.0
attrs==22.2.0
packaging==24.2
typing_extensions==4.12.2
zipp==3.21.0
platformdirs==4.3.6
requests==2.32.3
charset-normalizer==3.4.1
idna==3.10
certifi==2025.1.31

# Machine Learning
scikit_learn==1.2.2
tensorflow_intel==2.10.0
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-model-optimization==0.8.0
tensorflow-io-gcs-filesystem==0.31.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
xgboost==2.1.4
scikeras==0.10.0
hmmlearn==0.2.8
lightgbm==3.3.5
dm-tree==0.1.9
keras==2.10.0
Keras-Preprocessing==1.1.2
statsmodels==0.13.5
scipy==1.10.1
scikit-image==0.19.3
transformers==4.49.0

# Feature Explanation and Interpretation
shap==0.46.0
lime==0.2.0.1
eli5==0.13.0
alibi==0.9.0

# Time Series and Finance
yfinance==0.2.54
pandas_market_calendars==4.1.4
exchange_calendars==4.9
finnhub_python==2.4.22
pyluach==2.2.0
korean-lunar-calendar==0.3.1
multitasking==0.0.11
frozendict==2.4.6

# Plotting and Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.13.1
altair==4.2.0
pydeck==0.9.1
contourpy==1.3.1
graphviz==0.20.3
kiwisolver==1.4.8
pillow==9.5.0
cycler==0.12.1
fonttools==4.56.0
pyparsing==3.2.1

# Streamlit
streamlit==1.12.0
watchdog==6.0.0
toml==0.10.2
protobuf==3.19.6
blinker==1.9.0
validators==0.34.0
tornado==6.4.2
Pympler==1.1
toolz==1.0.0
semver==3.0.4
rich==13.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
cachetools==5.5.2
pyarrow==19.0.1
tenacity==9.0.0

# Scheduling
APScheduler==3.11.0
tzlocal==5.3
tzdata==2025.1
pytz==2025.1

# Wavelet
PyWavelets==1.4.1

# API
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
MarkupSafe==3.0.2
click==8.1.8

# Crypto
pycoingecko==3.2.0

# Optimization
optuna==3.2.0
alembic==1.14.1
SQLAlchemy==2.0.38
cmaes==0.11.1
greenlet==3.1.1
Mako==1.3.9

# Acceleration
numba==0.61.0
llvmlite==0.44.0

# Development and Testing
pytest==8.3.5
black==25.1.0
flake8==7.1.2
pylint==3.3.4
ruff==0.9.9
isort==6.0.1
mypy-extensions==1.0.0
absolufy-imports==0.3.1
vulture==2.14
pycodestyle==2.12.1
autoflake==2.3.1
pyflakes==3.2.0
astroid==3.3.8
mccabe==0.7.0
tomli==2.2.1
dill==0.3.9
pathspec==0.12.1
tomlkit==0.13.2
pluggy==1.5.0
exceptiongroup==1.2.2
iniconfig==2.0.0
pycln==2.5.0
bandit==1.8.3
pbr==6.1.1
stevedore==5.4.1
colorlog==6.9.0
gitdb==4.0.12
GitPython==3.1.44
smmap==5.0.2
pipreqs==0.5.0
yarg==0.1.9

# Data Processing
beautifulsoup4==4.13.3
soupsieve==2.6
lxml==5.3.1
html5lib==1.1
cloudpickle==3.1.1
fsspec==2025.2.0
tifffile==2025.2.18
imageio==2.37.0
networkx==3.4.2
cffi==1.17.1
pycparser==2.22
webencodings==0.5.1

# NLP Libraries
langcodes==3.5.0
language_data==1.3.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
regex==2024.11.6
tokenizers==0.21.0
safetensors==0.5.3
huggingface-hub==0.29.1
thinc==8.2.5
typer==0.15.2
confection==0.1.5
blis==0.7.11
catalogue==2.0.10
preshed==3.0.9
murmurhash==1.0.12
cymem==2.0.11
wasabi==1.1.3
srsly==2.5.1
smart-open==7.1.0
shellingham==1.5.4
weasel==0.4.1
cloudpathlib==0.20.0
annotated-types==0.7.0
pydantic==2.10.6
pydantic_core==2.27.2

# Jupyter and Development
ipykernel==6.25.0
ipython==8.12.3
matplotlib-inline==0.1.7
comm==0.2.2
debugpy==1.8.12
jupyter_client==8.6.3
jupyter_core==5.7.2
pyzmq==26.2.1
traitlets==5.14.3
decorator==5.2.1
asttokens==3.0.0
executing==2.2.0
pure_eval==0.2.3
stack-data==0.6.3
backcall==0.2.0
pickleshare==0.7.5
prompt_toolkit==3.0.50
jedi==0.19.2
parso==0.8.4
bleach==6.2.0
nbconvert==7.16.6
nbformat==5.10.4
jupyterlab_pygments==0.3.0
mistune==3.1.2
tinycss2==1.4.0
nbclient==0.10.2
defusedxml==0.7.1
fastjsonschema==2.21.1
pandocfilters==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
referencing==0.36.2
rpds-py==0.23.1
entrypoints==0.4
nest-asyncio==1.6.0

# TensorFlow and ML Dependencies
absl-py==1.4.0
astunparse==1.6.3
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.8.0
libclang==18.1.1
ml_dtypes==0.5.1
oauthlib==3.2.2
opt_einsum==3.4.0
requests-oauthlib==2.0.0
rsa==4.9
termcolor==2.5.0
threadpoolctl==3.5.0
wrapt==1.17.2
slicer==0.0.8
patsy==1.0.1

# Other Utilities
appdirs==1.4.4
colorama==0.4.6
docopt==0.6.2
filelock==3.17.0
libcst==1.6.0
namex==0.0.8
narwhals==1.28.0
optree==0.14.1
peewee==3.17.9
psutil==5.9.5
python-dateutil==2.9.0.post0
pywin32==308
tabulate==0.9.0
urllib3==1.26.20
vega-datasets==0.9.0
Markdown==3.7
cryptography==44.0.2
importlib_metadata==8.6.1
